+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.2.0-dev
 
 - Update to `dart_prism` 0.2.0.
+- **Breaking change**: Remove `root` from `PrismStyle`.
+- **Breaking change**: Change return type of `Prism.render` from `TextSpan` to
+  List<TextSpan>.
 
 ## 0.1.1